requirements.txt
flask
Flask-Login
python-dotenv
tortoise-orm
aiohttp